# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Every non-2xx response from `/api` has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MeetingCreate": {
        "properties": {
          "expected_person_ids": {
            "description": "Who you expect to attend. Pre-seeds speaker matching, which is more reliable than matching on platform display names alone.",
            "items": {
              "type": "string"
            },
            "title": "Expected Person Ids",
            "type": "array"
          },
          "meeting_url": {
            "description": "Google Meet URL. Meet AGI dispatches a bot to join.",
            "examples": [
              "https://meet.google.com/abc-defg-hij"
            ],
            "minLength": 8,
            "title": "Meeting Url",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "meeting_url"
        ],
        "title": "MeetingCreate",
        "type": "object"
      },
...
      "MeetingState": {
        "enum": [
          "scheduled",
          "joining",
          "in_call",
          "ended",
          "failed"
        ],
        "title": "MeetingState",
        "type": "string"
      },
...
      "Page_Meeting_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Meeting"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[Meeting]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for **Meet AGI**, a Google Meet copilot that actually speaks.\n\n### Building the frontend against this\n\nYou do not need a Google Meet, a Recall.ai key, or an internet connection. Start a\nfixture replay and subscribe to the socket:\n\n```bash\ncurl -X POST localhost:8000/api/dev/harness/start \\\n  -H 'content-type: application/json' \\\n  -d '{\"fixture_id\":\"q3_revenue_review\",\"speed\":6,\"loop\":true}'\n```\n\nThat emits the complete real-time event stream \u2014 participants, partial and final\ntranscript, speaking states, wake-word detection, agent state transitions, and\ninterjections with real citations. It is byte-for-byte the same contract a real meeting\nproduces; only `Meeting.source` differs.\n\n### Contract rules\n\n- Types are generated from this OpenAPI document. Never hand-write an API type.\n- All mutations are REST. The WebSocket is server-to-client only.\n- Adding fields, enum variants, and event types is free and unannounced. Renames and\n  removals get a heads-up first.\n- `switch` on `LiveEvent.type` with a `default` branch. New event types will arrive.\n",
    "title": "Meet AGI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meetings": {
      "get": {
        "operationId": "list_meetings_api_meetings_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeetingState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Meeting_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List meetings",
        "tags": [
          "meetings"
        ]
      },
      "post": {
        "description": "Dispatches a Recall.ai bot and returns immediately with the meeting in `joining`. The bot is not audible until someone admits it from the Google Meet waiting room; watch for `state: in_call` on the socket, or poll this meeting.\n\nThe bot can speak from the moment it is admitted \u2014 see `POST /api/meetings/{id}/speak`. Transcript ingestion lands separately.\n\nWithout `RECALL_API_KEY` set, this returns a `failed` meeting explaining how to run a fixture-backed meeting instead. Nothing crashes; the frontend can build both paths.",
        "operationId": "create_meeting_api_meetings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Send Meet AGI to a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
    "/api/meetings/{meeting_id}": {
      "delete": {
        "description": "Removes the session and its archived bundle from disk. Irreversible. A live session is stopped first.",
        "operationId": "delete_meeting_api_meetings__meeting_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Delete a session",
        "tags": [
          "review"
        ]
      },
      "get": {
        "operationId": "get_meeting_api_meetings__meeting_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
    "/api/meetings/{meeting_id}/ask": {
      "post": {
        "description": "Type a question from the dashboard. With `speak: true` Meet AGI also says the answer into the meeting.\n\nThis runs the same retrieval and reasoning that the spoken wake word does, so it is a genuine substitute rather than a demo prop: if wake detection misfires on stage, type the question and Meet AGI still answers correctly, out loud.\n\nWith no `ANTHROPIC_API_KEY` configured it returns a placeholder answer so the card, citations, and speaking flow can still be built against it.",
        "operationId": "ask_api_meetings__meeting_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Ask Meet AGI a question directly",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/api/meetings/{meeting_id}/interjections": {
      "get": {
        "operationId": "get_interjections_api_meetings__meeting_id__interjections_get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Interjection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List interjections for a meeting",
        "tags": [
          "meetings"
        ]
      }
    },
...
    "/api/meetings/{meeting_id}/leave": {
      "post": {
        "operationId": "leave_meeting_api_meetings__meeting_id__leave_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Make Meet AGI leave",
        "tags": [
          "meetings"
        ]
      }
    },
    "/api/meetings/{meeting_id}/mute": {
      "post": {
        "description": "Hard override. While muted Meet AGI will not post to chat or speak.",
        "operationId": "mute_api_meetings__meeting_id__mute_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Mute or unmute Meet AGI",
        "tags": [
          "meetings"
        ]
      }
    },
...
  }
}